fix: SharedMemory real + intro rock 70s + IA + monitor élite (resuelto conflicto)

+++ app/src/main/cpp/shm_hyperplane.cpp
@@ -4,7 +4,7 @@
 
 extern "C" JNIEXPORT jint JNICALL
 Java_com_ivannafusion_ShmManager_memfdCreate(JNIEnv*, jobject, jstring, jint) {
-    return -1; // Ya no se usa
+    return -1;
 }
 
 extern "C" JNIEXPORT jint JNICALL
@@ -14,5 +14,5 @@ Java_com_ivannafusion_ShmManager_nativeFtruncate(JNIEnv*, jobject, jint, jlong) 
 
 extern "C" JNIEXPORT jint JNICALL
 Java_com_ivannafusion_ShmManager_nativeMlock(JNIEnv*, jobject, jlong addr, jlong len) {
-    return mlock(reinterpret_cast<void*>(addr), len);
+    return mlock(reinterpret_cast<void*>(addr), (size_t)len);
 }
